%<d-% {pryr}	R Documentation

Create an delayed binding.

Description

Infix form of delayedAssign which creates an delayed or lazy binding, which only evaluates the expression the first time it is used.

Usage

x %<d-% value

Arguments

x
unquoted expression naming variable to create

value
unquoted expression to evaluate the first time name is accessed

Examples

x %<d-% (a + b)
a <- 10
b <- 100
x

[Package pryr version 0.1.4 Index]