call_tree {pryr}	R Documentation

Display a call (or expression) as a tree.

Description

call_tree takes a quoted expression. ast does the quoting for you.

Usage

call_tree(x, width = getOption("width"))

ast(x)

Arguments

x
quoted call, list of calls, or expression to display

width
displays width, defaults to current width as reported by getOption("width")

Examples

call_tree(quote(f(x, 1, g(), h(i()))))
call_tree(quote(if (TRUE) 3 else 4))
call_tree(expression(1, 2, 3))

ast(f(x, 1, g(), h(i())))
ast(if (TRUE) 3 else 4)
ast(function(a = 1, b = 2) {a + b})
ast(f()()())

[Package pryr version 0.1.4 Index]