names_c {pryr}	R Documentation

Extract function table from names.c from R subversion repository.

Description

Since this is an expensive operation, it is done once and cached within a session.

Usage

names_c()

Value

A data frame with columns

name
the function name in R

c-entry
The name of the corresponding C function, actually declared in ../include/Internal.h. All start with "do_", return SEXP, and have argument list (SEXP call, SEXP op, SEXP args, SEXP env)

offset
the 'op' (offset pointer) above; used for C functions which deal with more than one R function

eval
XYZ (three digits)

X=0 says that we should force R_Visible on
X=1 says that we should force R_Visible off
X=2 says that we should switch R_Visible on but let the C code update it.

Y=1 says that this is an internal function which must be accessed with a .Internal(.) call, any other value is accessible directly and printed in R as ".Primitive(..)".

Z=0 says don't evaluate (SPECIALSXP).
Z=1 says evaluate arguments before calling (BUILTINSXP)

arity
How many arguments are required/allowed; "-1" meaning “any”

pp-kind
Deparsing Info (-> PPkind in ../include/Defn.h )

precedence
Operator precedence (-> PPprec in ../include/Defn.h )

rightassoc
Right or left associative operator

[Package pryr version 0.1.6 Index]