rls {pryr}	R Documentation

Recursive ls.

Description

Performs ls all the way up to a top-level environment (either the parent of the global environment, the empty environment or a namespace environment).

Usage

rls(env = parent.frame(), all.names = TRUE)

Arguments

env
environment to start the search at. Defaults to the parent.frame. If a function is supplied, uses the environment associated with the function.

all.names
Show all names, even those starting with .? Defaults to TRUE, the opposite of ls

Author(s)

Winston Chang

[Package pryr version 0.1.4 Index]