substitute_q {pryr}	R Documentation

A version of substitute that evaluates its first argument.

Description

This version of substitute is needed because substitute does not evaluate it's first argument, and it's often useful to be able to modify a quoted call.

Usage

substitute_q(x, env)

Arguments

x
a quoted call

env
an environment, or something that behaves like an environment (like a list or data frame), or a reference to an environment (like a positive integer or name, see as.environment for more details)

Examples

x <- quote(a + b)
substitute(x, list(a = 1, b = 2))
substitute_q(x, list(a = 1, b = 2))

[Package pryr version 0.1.4 Index]