rebind {pryr}	R Documentation

Rebind an existing name.

Description

This function is similar to <<- with two exceptions:

Usage

rebind(name, value, env = parent.frame())

Arguments

name
name of existing binding to re-assign

value
new value

env
environment to start search in.

Details

if no existing binding is found, it throws an error

it does not recurse past the global environment into the attached packages

Examples

a <- 1
rebind("a", 2)
a
# Throws error if no existing binding
## Not run: rebind("b", 2)

local({
  rebind("a", 3)
})
a

# Can't find get because doesn't look past globalenv
## Not run: rebind("get", 1)

[Package pryr version 0.1.6 Index]